requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.2
jinja2==3.1.2
cryptography==41.0.7

# Phase 1 — Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Phase 2 — Rate Limiting
slowapi==0.1.9

# Phase 3 — Multi-channel Notifications
httpx==0.27.0

# Phase 4 — AI Features
scikit-learn==1.4.2
numpy==1.26.4
google-generativeai==0.5.4
